Make knowledge modules file-modules
