+++ packages/graph/hash_graph/lib/graph/src/knowledge/entity.rs
@@ -1,0 +1,98 @@
+use std::{collections::HashMap, fmt};
+
+use serde::{Deserialize, Serialize};
+use tokio_postgres::types::{FromSql, ToSql};
+use type_system::uri::BaseUri;
+use utoipa::ToSchema;
+use uuid::Uuid;
+
+#[derive(
+    Debug, Copy, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, ToSchema, FromSql, ToSql,
+)]
+#[repr(transparent)]
+#[postgres(transparent)]
+pub struct EntityId(Uuid);
+
+impl EntityId {
+    #[must_use]
+    pub const fn new(uuid: Uuid) -> Self {
+        Self(uuid)
+    }
+}
+
+impl fmt::Display for EntityId {
+    fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(fmt, "{}", &self.0)
+    }
+}
+
+/// An entity.
+///
+/// When expressed as JSON, this should validate against its respective entity type(s).
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, ToSchema)]
+#[schema(value_type = Object)]
+pub struct Entity(HashMap<BaseUri, serde_json::Value>);
+
+impl Entity {
+    #[must_use]
+    pub const fn properties(&self) -> &HashMap<BaseUri, serde_json::Value> {
+        &self.0
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    fn test_entity(json: &str) {
+        let json_value: serde_json::Value = serde_json::from_str(json).expect("invalid JSON");
+
+        let entity: Entity = serde_json::from_value(json_value.clone()).expect("invalid entity");
+
+        assert_eq!(
+            serde_json::to_value(entity.clone()).expect("could not serialize"),
+            json_value,
+            "{entity:#?}"
+        );
+    }
+
+    #[test]
+    fn book() {
+        test_entity(graph_test_data::entity::BOOK_V1);
+    }
+
+    #[test]
+    fn address() {
+        test_entity(graph_test_data::entity::ADDRESS_V1);
+    }
+
+    #[test]
+    fn organization() {
+        test_entity(graph_test_data::entity::ORGANIZATION_V1);
+    }
+
+    #[test]
+    fn building() {
+        test_entity(graph_test_data::entity::BUILDING_V1);
+    }
+
+    #[test]
+    fn person() {
+        test_entity(graph_test_data::entity::PERSON_A_V1);
+    }
+
+    #[test]
+    fn playlist() {
+        test_entity(graph_test_data::entity::PLAYLIST_V1);
+    }
+
+    #[test]
+    fn song() {
+        test_entity(graph_test_data::entity::SONG_V1);
+    }
+
+    #[test]
+    fn page() {
+        test_entity(graph_test_data::entity::PAGE_V1);
+    }
+}
