+++ packages/graph/hash_graph/lib/graph/src/knowledge/link.rs
@@ -1,0 +1,66 @@
+use std::fmt;
+
+use serde::{Deserialize, Serialize};
+use type_system::uri::VersionedUri;
+use utoipa::ToSchema;
+
+use super::EntityId;
+
+/// A Link between a source and a target entity identified by [`EntityId`]s.
+///
+/// The link is described by a link type [`VersionedUri`].
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, ToSchema)]
+#[serde(rename_all = "camelCase")]
+pub struct Link {
+    source_entity_id: EntityId,
+    target_entity_id: EntityId,
+    #[schema(value_type = String)]
+    link_type_id: VersionedUri,
+    // TODO: Consider if ordering should be exposed on links as they are here. The API consumer
+    //   manages indexes currently.
+    //   https://app.asana.com/0/1202805690238892/1202937382769278/f
+    index: Option<i32>,
+}
+
+impl Link {
+    #[must_use]
+    pub const fn new(
+        source_entity_id: EntityId,
+        target_entity_id: EntityId,
+        link_type_id: VersionedUri,
+        index: Option<i32>,
+    ) -> Self {
+        Self {
+            source_entity_id,
+            target_entity_id,
+            link_type_id,
+            index,
+        }
+    }
+
+    #[must_use]
+    pub const fn source_entity(&self) -> EntityId {
+        self.source_entity_id
+    }
+
+    #[must_use]
+    pub const fn target_entity(&self) -> EntityId {
+        self.target_entity_id
+    }
+
+    #[must_use]
+    pub const fn link_type_id(&self) -> &VersionedUri {
+        &self.link_type_id
+    }
+
+    #[must_use]
+    pub const fn index(&self) -> Option<i32> {
+        self.index
+    }
+}
+
+impl fmt::Display for Link {
+    fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(fmt, "{:?}", &self)
+    }
+}
